EmailService 서비스 추상화 삭제
